// src/app/shared/types/api-types/bidt-qualif-employee-dto.interface.ts
export interface BidtQualifEmployeeDTO {
  bidtEmployeeId?: number;
  bidtQualificationId?: number;
  qualifEmployeeInstitution?: string;
  qualifEmployeeStartDate?: string;
  qualifEmployeeEndDate?: string;
  qualifOngoing?: boolean;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}